backend/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
langchain==1.1.3
langchain-text-splitters==1.0.0
langchain-openai==1.1.2
langchain-community==0.4.1
langgraph==1.0.4
openai==2.11.0
faiss-cpu==1.13.1
pymongo==4.15.5
pypdf2==3.0.1
python-docx==1.2.0
pydantic==2.12.5
pydantic-settings==2.12.0